% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnet15_utils.R
\name{BerkeleyJulianDateToPOSIXct}
\alias{BerkeleyJulianDateToPOSIXct}
\title{convert JulianDate format used in Berkeley release to POSIXct}
\usage{
BerkeleyJulianDateToPOSIXct(julianDate, tz = "UTC", ...)
}
\arguments{
\item{julianDate}{numeric vector representing times (see details for format)}

\item{tz}{time zone used to represent the dates}

\item{...}{further arguments to \code{\link{strptime}}}
}
\description{
convert JulianDate format used in Berkeley release to POSIXct
}
\details{
In the Berkeley-Release of the Fluxnet data, the time is stored as an number
with base10-digits representing YYYYMMddhhmm
}
\seealso{
\code{\link{POSIXctToBerkeleyJulianDate}}
}
