% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_cswi}
\alias{compute_cswi}
\title{Compute the conservative soil moisture index (CSWI)}
\usage{
compute_cswi(data, smax = 5)
}
\arguments{
\item{data}{data.frame with numeric columns \code{precip} and \code{ET} in mm}

\item{smax}{maximum water storage capacity in mm}
}
\value{
numeric vector of CSWI in mm
}
\description{
Compute the conservative soil moisture index (CSWI)
}
