% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_DWCI}
\alias{compute_DWCI}
\title{Diurnal water:carbon index (DWCI)}
\usage{
compute_DWCI(data, units_per_day = 48)
}
\arguments{
\item{data}{data.frame of sub-daily timeseries with variables
\describe{
\item{Rg_pot}{Potential radiation}
\item{LE}{evapotranspiration or latent energy}
\item{GPP}{ross primary productivity}
\item{VPD}{vapor pressure deficit}
\item{NEE}{net ecosystem exchange}
\item{LE_sd}{estimation of the uncertainty of LE}
\item{GPP_sd}{eestimation of the uncertainty of GPP}
\item{NEE_fall}{ Modeled net ecosystem exchange i.e. no noise}
\item{LE_fall}{Modeled evapotranspiration or latent energy i.e. no noise}
}}

\item{units_per_day}{integer: frequency of the sub-daily measurements,
48 for half hourly measurements'}
}
\value{
The diurnal water:carbon index (DWCI)
}
\description{
DWCI measures the probability that the carbon and water are coupled
within a given day. Method takes the correlation between
evapotranspiration (LE) and gross primary productivity (GPP) and
calculates the correlation within each day. This correlation is then
compared to a distribution of correlations between artificial datasets
built from the signal of potential radiation and the uncertainty in
the LE and GPP.
}
