% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_GPPgrad}
\alias{compute_GPPgrad}
\title{Compute the seasonal gradiant from smoothed daily GPP.}
\usage{
compute_GPPgrad(GPP, units_per_day = 48, sigma = 20)
}
\arguments{
\item{GPP}{numeric vector of gross primary productivity (umol m-2 s-1)}

\item{nStepsPerDay}{number of records per day}
}
\value{
numeric vector of gradients
}
\description{
Compute the seasonal gradiant from smoothed daily GPP.
}
