% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_norm_diurnal_centroid}
\alias{compute_norm_diurnal_centroid}
\title{Normalized diurnal centroid of latent energy (LE)}
\usage{
compute_norm_diurnal_centroid(flux, Rg, nrecday = 48)
}
\arguments{
\item{flux}{numeric vector of sub-daily flux (usuall LE)
that must be condinuous and regular}

\item{Rg}{incoming radiation, can be any unit}

\item{nrecday}{integer: frequency of the sub-daily measurements,
48 for half hourly measurements}
}
\value{
The normalized diurnal centroid, in hours,
at a daily frequency
}
\description{
Calculates the diurnal centroid of LE relative to the diurnal centroid of
incoming radiation (Rg).
}
