% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_simplifiedDWCI}
\alias{compute_simplifiedDWCI}
\title{simplified Diurnal water:carbon index (DWCI)}
\usage{
compute_simplifiedDWCI(data, nrecday = 48, na_value = 0)
}
\arguments{
\item{data}{data.frame of sub-daily timeseries with variables
\describe{
\item{Rg_pot}{Potential radiation}
\item{ET}{evapotranspiration or latent energy}
\item{GPP}{ross primary productivity}
\item{VPD}{vapor pressure deficit}
\item{NEE}{net ecosystem exchange}
\item{ET_sd}{estimation of the uncertainty of ET}
\item{GPP_sd}{eestimation of the uncertainty of GPP}
}}

\item{nrecday}{integer: frequency of the sub-daily measurements,
48 for half hourly measurements'}

\item{na_value}{numeric scalar: to replace NA values in correlation
Defaults to 0.0 - i.e. no probability of correlation in DWCI.}
}
\value{
numeric vector length(data)/nrecday: diurnal water:carbon index (DWCI)
}
\description{
similar to DWCI it
measures the probability that the carbon and water are coupled
within a given day. In difference to the full DWCI the correlation
only the standard deviation of ET and GPP is used but not the
daily correlation between NEE and ET errors.
Hence, it can be used when noisefree, i.e. modelled, NEE_fall and ET_fall
are not available.
}
