% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priego.R
\name{compute_stomatal_conductance_jarvis}
\alias{compute_stomatal_conductance_jarvis}
\title{Calculate stomatal conductance using Jarvis's approach.}
\usage{
compute_stomatal_conductance_jarvis(par, Q, VPD, Tair, gcmax)
}
\arguments{
\item{par}{Set of parameter for the respective sensitivity function.}

\item{Q}{Vector containing time series of
photosynthetic active radiation (umol CO2 m-2 s-1)}

\item{VPD}{Vector containing time series of
vapor pressure deficit (kPa).}

\item{Tair}{Vector containing time series of
air temperature (deg C).}

\item{gcmax}{Empirical parameter defining
maximum conductance (m s-1 or mol m-2 s-1).}
}
\value{
a numeric value:
\item{compute_stomatal_conductance_jarvis}{canopy leaf conductance (untis refer to that given by gmax)}
}
\description{
Calculate canopy stomatal conductance using Jarvis's approach.
}
\details{
the following metrics are calculated:\preformatted{       \deqn{compute_stomatal_conductance_jarvis <- gcmax*f(Q)*f(VPD)*f(Tair)}
}
}
\examples{
 ## Selecting a single day (e.g. 15-05-2011)
 tmp <-  EddySample[ EddySample$TIMESTAMP_START>  201105150000,]
 tmp <-  tmp[tmp$TIMESTAMP_START<  201105160000,]
compute_stomatal_conductance_jarvis(par=c(200, 0.2, 25)
,Q = tmp$PPFD_IN * 2 #(convert to mumuol)
,VPD = tmp$VPD_F
,Tair = tmp$TA_F
,gcmax = 1)
}
\references{
Perez-Priego, O., G. Katul, M. Reichstein et al. Partitioning eddy covariance
water flux components using physiological and micrometeorological approaches,
Journal of Geophysical Research: Biogeosciences. In press
}
