% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_TandE}
\alias{compute_TandE}
\title{compute T and E from given ET and quantiles of water use efficiency (WUE)}
\usage{
compute_TandE(ET, wue_pred)
}
\arguments{
\item{ET}{numeric vector of evapotranspiration}

\item{wue_pred}{tibble with each column a prediction quantile of WUE
specifying the quantile as a number (0.0..1.0) in the column name}
}
\value{
tibble in long format with columsn ET, perc (1..100), WUE, T, E
column id identifies the original row-number in the wide format
}
\description{
compute T and E from given ET and quantiles of water use efficiency (WUE)
}
