% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnet15_utils.R
\name{construct_data_FN15}
\alias{construct_data_FN15}
\title{Format Fluxnet2015 data.frame}
\usage{
construct_data_FN15(
  data,
  GPP = "GPP_NT_VUT_USTAR50",
  GPP_sd = "NEE_VUT_USTAR50_RANDUNC",
  Ca = "CO2_F_MDS",
  H = "H_F_MDS",
  LE = "LE_F_MDS",
  LE_QC = "LE_F_MDS_QC",
  LE_sd = "LE_RANDUNC",
  NEE_QC = "NEE_VUT_USTAR50_QC",
  NIGHT = "NIGHT",
  precip = "P",
  Pair = "PA",
  Q = "PPFD_IN",
  Rg = "SW_IN_F_MDS",
  Rg_pot = "SW_IN_POT",
  Tair = "TA_F_MDS",
  ustar = "USTAR",
  VPD = "VPD_F_MDS",
  u = "WS_F"
)
}
\arguments{
\item{data}{data frame with column names given by the following arguments.}

\item{GPP}{photosynthesis data (umol CO2 m-2 s-1).}

\item{GPP_sd}{photosynthesis uncertainties (umol CO2 m-2 s-1).}

\item{Ca}{atmospheric CO2 concentration (umol Co2 mol air-1).}

\item{H}{sensible heat flux (W m-2).}

\item{LE}{Latent heat flux (W m-2)}

\item{LE_QC}{quality flag for LE (0 measured, 1 good quality, ...)}

\item{LE_sd}{uncertainty of LE}

\item{NEE_QC}{quality flag for NEE (0 measured, 1 good quality, ...)}

\item{NIGHT}{Flag indicating nighttime interval (1 for night, zero otherwise)}

\item{precip}{P used if measured (mm per dataset resolution: either hour or half-hour)}

\item{Pair}{atmospheric pressure (kPa).}

\item{Q}{photosynthetic active radiation (umol m-2 s-1).}

\item{Rg}{Shortwave radiation, incoming (W/m2)}

\item{Rg_pot}{potential Shortwave radiation, incoming (W/m2)}

\item{Tair}{air temperature (deg C).}

\item{ustar}{wind friction velocity (m s-1).}

\item{VPD}{vapor pressure deficit (hPa).}

\item{u}{wind velocity (m s-1).}
}
\value{
data.frame with colums as given by argument names and
\item{timestamp}{POSIXct time: end of averaging period}
\item{ET}{evapotranspiration computed from LE in mm/hour}
\item{ET_sd}{uncertainty of ET}
\item{RH}{relative humidity (1/1) computed from VPD and Tair}
\item{isnight}{boolean TRUE if its nighttime computed from NIGHT}
\item{qualtiy_flag}{
(NEE_QC == 0) & (LE_QC == 0) & is.finite(ET) & is.finite(GPP)}
}
\description{
Construct data.frame formatted for partitioning from Fluxnet2015 format
https://fluxnet.org/data/fluxnet2015-dataset/
}
