% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{correlations_from_artificial}
\alias{correlations_from_artificial}
\title{Compute Artificial correlations from daily data}
\usage{
correlations_from_artificial(dfs, nboot = 100)
}
\arguments{
\item{dfs}{tibble with columns GPP, ET, GPP_sd, ET_sd, NEE_err, ET_err}
}
\value{
data.frame by day with columns mean_GPP, mean_ET,
corr_err (pearson correlation coefficient between errors of ET and NEE),
corr_syn (vector of perason correlation coefficients between GPP and ET)
dwci (rank of observed correlation within artificial correlations in \%)
}
\description{
Correlation between ET and GPP is obscured by noise. The effect size
depends on signal (flux) to noise (errors) ratio.
This function bootstaps the correlation of two perfectly correlated signals
with random noise added corresponding to errors in ET and NEE.
and then gives the rank
}
