% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{daily_corr}
\alias{daily_corr}
\title{Daily correlation coefficient}
\usage{
daily_corr(x, y, Rg_pot, units_per_day = 48)
}
\arguments{
\item{Rg_pot}{potential incoming radiation}

\item{units_per_day}{integer: frequency of the sub-daily measurements,
48 for half hourly measurements}
}
\value{
correlation coefficents at daily timescale
}
\description{
Calculates a daily correlation coefficient between two sub-daily timeseries
\code{x} and \code{y} considering only daytime, i.e. \code{Rg > 0}.
Vectors \code{x}, \code{y}, and \code{Rg} must have the same length.
Only complete cases, i.e. no NAs are considered.
}
