% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETha.R
\docType{data}
\name{DETha}
\alias{DETha}
\title{Example of eddy covariance data.}
\format{
Corresponding to return of  \code{\link{construct_data_FN15}}
}
\source{
\url{http://fluxnet.fluxdata.org/data/fluxnet2015-dataset/}
}
\usage{
DETha
}
\description{
A dataset containing eddy covariance data and meteorological data collected
at the Tharandt site.
}
\details{
TODO
}
\keyword{datasets}
