% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priego.R
\name{estimate_T_priego_5days}
\alias{estimate_T_priego_5days}
\title{Estiamte parameters for 5 day windows around iday and predict for iday}
\usage{
estimate_T_priego_5days(
  data,
  iday,
  chi_o,
  WUE_o,
  config = priego_config(),
  par_start = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame entire dataset}

\item{iday}{cumday to process}

\item{chi_o}{long-term internal leaf to ambient CO2 concentration}

\item{WUE_o}{long-term water use efficiency}

\item{config}{configuration passed to \code{optim_priego}}

\item{par_start}{starting parameters, may pass from last day for efficiency}

\item{...}{passed to optim_priego and predict_transpiration_opriego}
}
\value{
list with entries
\item{data}{data.frame for given day with added/updated columns
T (mm/hr), T_sd, GPP_pred, and GPP_pred_sd}
\item{paropt}{numeric vector of mean parameters (not the best estimates)}
}
\description{
Estiamte parameters for 5 day windows around iday and predict for iday
}
