% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{etpart_constants}
\alias{etpart_constants}
\title{Constants used throughout ETPart}
\usage{
etpart_constants(Cp = 1003.5, R_gas_constant = 0.287058, M_air = 0.0289644)
}
\arguments{
\item{Cp}{heat capacity (J kg-1 K-1)}

\item{R_gas_constant}{(J kg-1 deg K-1)}

\item{M_air}{molar mass of air (kg mol-1)}
}
\value{
a list with above arguments as entries.
}
\description{
Constants used throughout ETPart
}
