% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIHyy.R
\docType{data}
\name{FIHyy}
\alias{FIHyy}
\title{Example of eddy covariance data.}
\format{
Corresponding to return of  \code{\link{construct_data_FN15}}
}
\source{
\url{http://fluxnet.fluxdata.org/data/fluxnet2015-dataset/}
}
\usage{
FIHyy
}
\description{
A dataset containing eddy covariance data and meteorological data collected
at the Hyytiala site.
Data is included in the FLUXNET2015 dataset.
Adopted from https://github.com/jnelson18/ecosystem-transpiration
}
\keyword{datasets}
