% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnet.R
\name{get_fluxnet_metadata}
\alias{get_fluxnet_metadata}
\title{Extracts the site metadata from fluxdatasite url}
\usage{
get_fluxnet_metadata(site, fluxdatasite = "http://sites.fluxdata.org")
}
\arguments{
\item{site}{site name, e.g. FI-Hyy}

\item{fluxdatasite}{url to be appended by /site}
}
\value{
list named by first entry in html-table and entries
contact.name: vector of strings, contact.email: vector of strings parsed
from entry Tower.Team
}
\description{
Extracts the site metadata from fluxdatasite url
}
