% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnet15_utils.R
\name{get_tzone}
\alias{get_tzone}
\title{Extract the timezone attribute from POSIXct with default on missing}
\usage{
get_tzone(x, default = "GMT")
}
\arguments{
\item{x}{POSIXct vector}

\item{default}{time zone returned, if x has not timezone associated or
attribute is the zero string}
}
\description{
Extract the timezone attribute from POSIXct with default on missing
}
\examples{
get_tzone(as.POSIXct("2010-07-01 16:00:00", tz = "etc/GMT-1") )
get_tzone(as.POSIXct("2010-07-01 16:00:00") )
# printed with local time zone, but actually has no tz attribute
get_tzone(Sys.time())

}
