% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priego.R
\name{partition_priego}
\alias{partition_priego}
\title{partition ET by Priego approach}
\usage{
partition_priego(data, control = priego_config(), ...)
}
\arguments{
\item{data}{data.frame with required columns: XX}

\item{control}{list with configuration options see \code{\link{priego_config}}}
}
\value{
see \code{\link{tea_predict}}, \code{data} with predictions
percentiles of WUE, E and T appended.
}
\description{
partition ET by Priego approach
}
