% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnet15_utils.R
\name{POSIXctToBerkeleyJulianDate}
\alias{POSIXctToBerkeleyJulianDate}
\title{convert POSIXct to JulianDate format used in Berkeley release}
\usage{
POSIXctToBerkeleyJulianDate(sDateTime, tz = get_tzone(sDateTime))
}
\arguments{
\item{sDateTime}{POSIXct vector}

\item{tz}{timezone to use}
}
\description{
convert POSIXct to JulianDate format used in Berkeley release
}
\details{
In the Berkeley-Release of the Fluxnet data, the time is stored as an number
with base10-digits representing YYYYMMddhhmm
}
\seealso{
\code{\link{BerkeleyJulianDateToPOSIXct}}
}
