% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{pred_ranger_quantiles}
\alias{pred_ranger_quantiles}
\title{Predict quantiles from workflow's ranger fitting object for new data}
\usage{
pred_ranger_quantiles(rf, newdata, quantiles = c(0.5, 0.75))
}
\arguments{
\item{rf}{tidymodels workflow}

\item{newdata}{data.frame with predictors as in traning data}

\item{quantiles}{numeric vector of probabilities of prediction distribution}
}
\value{
data.frame with one column for each quantile
}
\description{
The workflow need to be constructed with
\code{set_engine("ranger", quantreg = TRUE, ...)}.
The newdata is transformed/prepared by \code{\link{bake}}.
}
