% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{rep_daily_aggregate}
\alias{rep_daily_aggregate}
\title{Compute daily aggregate and repeat it for each record}
\usage{
rep_daily_aggregate(x, FUN, nrecday)
}
\arguments{
\item{x}{vector}

\item{FUN}{\code{function(x) -> numeric scalar} to aggregate over
each column of a matrix with a day in each row, i.e. \code{colSums} or
\code{applycols(sum)}
daily numeric vector}

\item{nrecday}{number of records fore each day}
}
\value{
vector of the same length as x
}
\description{
x must have the same number of records for each day and contain
only complete days.
}
