% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{tea_filter}
\alias{tea_filter}
\title{Filter conditions for training the TEA WUE model.}
\usage{
tea_filter(data, control)
}
\arguments{
\item{data}{data.frame with columns GPP, Tair, Rg, and cswi}

\item{control}{list with entries GPPlimit, GPPdaylimit, Tairlimit, Rglimit,
and CWSIlimit. See \code{\link{tea_config}}.}
}
\value{
data.frame containing only rows matching the filter criteria.
}
\description{
Filter conditions for training the TEA WUE model.
}
