% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{tea_predict}
\alias{tea_predict}
\title{Predict water use efficiency WUE, transpiration T, and evaporation E}
\usage{
tea_predict(rf, data, control)
}
\arguments{
\item{rf}{tidymodels workflow}

\item{data}{data.frame with predictors as in traning data}

\item{control}{see \code{\link{tea_config}}}
}
\value{
XXTODO
}
\description{
Predict water use efficiency WUE, transpiration T, and evaporation E
}
