% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{tea_preprocess}
\alias{tea_preprocess}
\title{preprocess for TEA partitioning}
\usage{
tea_preprocess(data, control = tea_config())
}
\arguments{
\item{data}{data.frame of full days, must contain at least timestamp,
ET, GPP, RH, Rg, Rg_pot, Tair, VPD, precip,}

\item{control}{}
}
\value{
ds with filters and additional indices
\itemize{
\item \code{cswi}: see \code{\link{compute_cswi}},
\item \code{C_ET}, \code{C_Rg}: diurnal centroids,
see \code{\link{compute_diurnal_centroid}},
\item \code{dcwi}: see \code{\link{compute_DWCI}},
\item \code{Rg_pot_daily}: daily daily Rg_pot in MJ m-2 d-1
\item \code{year}: the year of the time-stamp
use mid-time-stamps. If using end-timestampes, NewYear is already next
\item \code{GPPgrad}: daily smoothed GPP gradient in umol C m-2 s-1 d-1
, which gives and indication of phenology
\item \code{Rgpotgrad} and \code{Rpotgrad_day}: gradients in \code{Rpot}
and daily means of \code{Rpot}, which give an indication of time
in the year, ie. season
\item \code{tempFlag}: records with minimum temperature
(see \code{\link{tea_config}})
\item \code{GPPFlag}: records with minimum GPP and minimum daily GPP
(see \code{\link{tea_config}})
\item \code{seasonFlag}: combined \code{tempFlag} and \code{GPPFlag}
\item \code{inst_WUE}: instant water use efficiency (GPP/ET)
in g C per kg H2O
}
}
\description{
Builds all the derived variables used in the partitioning such as CSWI,
DWCI, etc., as well as filters which remove night time, low air temp/GPP,
etc. periods.
}
