% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_daily_GPP}
\alias{compute_daily_GPP}
\title{sum GPP by day of year}
\usage{
compute_daily_GPP(GPP, timestamp)
}
\arguments{
\item{GPP}{numeric vector to be summed}

\item{timestamp}{POSIXct vector of times}

\item{doy}{factor alternative to timestamp to specify the day of year}
}
\value{
numeric vector (length GPP) with corresponding daily sum of GPP
}
\description{
sum GPP by day of year
}
