% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{compute_GPPgrad}
\alias{compute_GPPgrad}
\title{Compute the seasonal gradiant from smoothed daily GPP.}
\usage{
compute_GPPgrad(GPP, nrecday = 48, sigma = 20)
}
\arguments{
\item{GPP}{numeric vector of gross primary productivity (umol m-2 s-1)}

\item{nrecday}{integer scalar: number of records within one day}

\item{sigma}{parameter to \code{\link{mmand::gaussianSmooth}} in units
number of days}
}
\value{
numeric vector of gradients, repeated nrecday to match length of GPP
}
\description{
Compute the seasonal gradiant from smoothed daily GPP.
}
