% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priego.R
\name{compute_stomatal_conductance_jarvis}
\alias{compute_stomatal_conductance_jarvis}
\title{Canopy stomatal conductance by Jarvis.}
\usage{
compute_stomatal_conductance_jarvis(par, Q, VPD, Tair, gcmax)
}
\arguments{
\item{par}{Set of parameter for the respective sensitivity function.}

\item{Q}{Vector containing time series of
photosynthetic active radiation (umol CO2 m-2 s-1)}

\item{VPD}{Vector containing time series of
vapor pressure deficit (kPa).}

\item{Tair}{Vector containing time series of
air temperature (deg C).}

\item{gcmax}{Empirical parameter defining
maximum conductance (m s-1 or mol m-2 s-1).}
}
\value{
a numeric value:
canopy leaf conductance (units refer to that given by gmax)
}
\description{
Calculate canopy stomatal conductance using Jarvis's approach.
}
\details{
the following metrics are calculated:

\deqn{compute_stomatal_conductance_jarvis <- gcmax*f(Q)*f(VPD)*f(Tair)}
}
\examples{
 ## Selecting a single day (e.g. 15-05-2010)
tz <- get_tzone(FIHyy$timestamp)
tmp <- FIHyy[(FIHyy$timestamp > ISOdatetime(2010,5,15,0,0,0,tz=tz)) &
             (FIHyy$timestamp <= ISOdatetime(2010,5,16,0,0,0,tz=tz)),]
gc = compute_stomatal_conductance_jarvis(par=c(200, 0.2, 25)
,Q = tmp$Q
,VPD = tmp$VPD/10 # convert hPa to kPa
,Tair = tmp$Tair
,gcmax = 1)
plot(gc ~ tmp$timestamp)
}
\references{
Perez-Priego, O., G. Katul, M. Reichstein et al. Partitioning eddy covariance
water flux components using physiological and micrometeorological approaches,
Journal of Geophysical Research: Biogeosciences. In press
}
\seealso{
\code{\link{partition_priego}},
}
