% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priego.R
\name{predict_transpiration_opriego}
\alias{predict_transpiration_opriego}
\title{Compute transpiration given optimized parameters}
\usage{
predict_transpiration_opriego(
  data,
  par,
  chi_o,
  WUE_o,
  config = priego_config(),
  ...
)
}
\arguments{
\item{data}{Data.frame with columns
\itemize{
\item GPP:     photosynthesis data (umol CO2 m-2 s-1).
\item GPP_sd:  photosynthesis uncertainties (umol CO2 m-2 s-1).
\item H:       sensible heat flux (W m-2).
\item VPD:     vapor pressure deficit (hPa).
\item Tair:    air temperature (deg C).
\item Pair:    atmospheric pressure (kPa).
\item Q:       photosynthetic active radiation (umol m-2 s-1).
\item Ca:      atmospheric CO2 concentration (umol Co2 mol air-1).
\item ustar:   wind friction velocity (m s-1).
\item u:       wind velocity (m s-1).
}}

\item{par}{optimized parameters see}

\item{chi_o}{Long-term effective chi}

\item{WUE_o}{Long-term effective WUE}

\item{...}{Further arguments to \code{predict_T_GPP_opriego}
such as a non-default \code{VPD_plant} or precomputed intermediates.}
}
\value{
vector of estimated transpiration for each record in data
?in mmol m-2 s-1
}
\description{
Compute transpiration given optimized parameters
}
