% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priego.R
\name{predict_transpiration_opriego}
\alias{predict_transpiration_opriego}
\title{Compute transpiration given parameters}
\usage{
predict_transpiration_opriego(data, par, chi_o, WUE_o, ...)
}
\arguments{
\item{data}{Data.frame with columns
\itemize{
\item GPP:     photosynthesis data (umol CO2 m-2 s-1).
\item GPP_sd:  photosynthesis uncertainties (umol CO2 m-2 s-1).
\item H:       sensible heat flux (W m-2).
\item VPD:     vapor pressure deficit (hPa).
\item Tair:    air temperature (deg C).
\item Pair:    atmospheric pressure (kPa).
\item Q:       photosynthetic active radiation (umol m-2 s-1).
\item Ca:      atmospheric CO2 concentration (umol Co2 mol air-1).
\item ustar:   wind friction velocity (m s-1).
\item u:       wind velocity (m s-1).
}}

\item{par}{optimized parameters see \code{\link{partition_priego}}}

\item{chi_o}{Long-term effective chi}

\item{WUE_o}{Long-term effective WUE}

\item{...}{Further arguments to \code{predict_T_GPP_opriego}
such as \code{constants} (see \code{\link{etpart_constants}})
or a non-default \code{VPD_plant} or
precomputed intermediates.}
}
\value{
vector of estimated transpiration for each record in data
in mm/hour (kg m-2 hour-1)
}
\description{
Compute transpiration given parameters
}
\seealso{
\code{\link{partition_priego}},
\code{\link{compute_stomatal_conductance_jarvis}}
}
