% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priego.R
\name{priego_config}
\alias{priego_config}
\title{Configure parameters of the Priego transpiration partitioning}
\usage{
priego_config(
  C = 1.189,
  par_lower = c(a1 = 0, Do = 0, To = 10, beta = 0),
  par_upper = c(a1 = 400, Do = 0.4, To = 30, beta = 1),
  niter = 20000,
  updatecov = 500,
  ntrydr = 3,
  burninlength = 10000,
  Cp = 1003.5,
  R_gas_constant = 0.287058,
  M_air = 0.0289644
)
}
\arguments{
\item{C}{Empirical coeficient for C3 species (see Wang et al., 2017; Plant Nature).}

\item{par_lower}{A vector containing the lower bound of the parameters
(a1,Do,To,beta)}

\item{par_upper}{A vector containing the upper bound of the parameters
(a1,Do,To,beta)}

\item{niter}{number of iterations for the MCMC}

\item{updatecov}{number of iterations after which the parameter covariance
matrix is (re)evaluated based on the parameters kept thus far, and used
to update the MCMC jumps.}

\item{ntrydr}{maximal number of tries for the delayed rejection procedure}

\item{burninlength}{number of initial iterations to be removed from output.}

\item{Cp}{heat capacity \link{J kg-1 K-1}}

\item{R_gas_constant}{\link{J kg-1 deg K-1}}

\item{M_air}{molar mass of air, \link{kg mol-1}}
}
\value{
a list with above arguments as entries.
}
\description{
Configure parameters of the Priego transpiration partitioning
}
