% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{tea_config}
\alias{tea_config}
\title{configure hyperparameters of the TEA algorithm}
\usage{
tea_config(CSWIlimit = -0.5, perc = c(0.75), smax = 5,
  GPPlimit = 0.05, GPPdaylimit = 0.5, Tairlimit = 5, Rglimit = 0)
}
\arguments{
\item{CSWIlimit}{filter for conservative surface
wetness index to be lower than this limit}

\item{perc}{numeric vector of percentiles of the prediction distribution
do average over}

\item{smax}{numeric scalar of maximum water storage
(see \code{\link{compute_cswi}})}

\item{GPPlimit}{numeric scalar (mumol/m2/s): filter for half-hours
with carbon fluxes larger than this half-hourly threshold}

\item{GPPdaylimit}{numeric scalar (mumol/m2/day): filter for days
with carbon fluxes larger than this daily threshold}

\item{Tairlimit}{numeric scalar (degree Celsius): filter for half-hours
with air temperatures larger than this threshold}

\item{Rglimit}{numeric scalar (W/m2): filter for half-hours
with incoming radiation larger than this threshold}

\item{s0}{numeric scalar of initial water storage
(see \code{\link{compute_cswi}})}
}
\value{
list with the arguments
}
\description{
configure hyperparameters of the TEA algorithm
}
