% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\name{tea_predict}
\alias{tea_predict}
\title{Predict water use efficiency WUE, transpiration T, and evaporation E}
\usage{
tea_predict(rf, data, control)
}
\arguments{
\item{rf}{tidymodels workflow}

\item{data}{data.frame with predictors as in traning data}

\item{control}{see \code{\link{tea_config}}}
}
\value{
\code{data} with appended columns: \code{WUE_perc}, \code{T_perc},
and \code{ET_perc}
where \code{perc} is each prediction percentile corresponding to the
\code{control$quantile_wue}, e.g. 75 for the 0.75 prediction quantile.
}
\description{
Predict water use efficiency WUE, transpiration T, and evaporation E
}
